depr: deprecate `utils.paths`
